Make ICompletionRules a concrete class with default implementations
