+++ src/Features/Core/Completion/AbstractCompletionService.cs
@@ -24,7 +24,7 @@ namespace Microsoft.CodeAnalysis.Completion
         private readonly List<string> _committedItems = new List<string>(MruSize);
         private readonly object _mruGate = new object();
 
-        private void CompletionItemCommitted(CompletionItem item)
+        internal void CompletionItemCommitted(CompletionItem item)
         {
             lock (_mruGate)
             {
@@ -40,7 +40,7 @@ namespace Microsoft.CodeAnalysis.Completion
             }
         }
 
-        protected int GetMRUIndex(CompletionItem item)
+        internal int GetMRUIndex(CompletionItem item)
         {
             lock (_mruGate)
             {
@@ -190,7 +190,7 @@ namespace Microsoft.CodeAnalysis.Completion
             return MergeAndPruneCompletionLists(allProvidersAndLists.Select(g => g.List), completionRules);
         }
 
-        private static CompletionList MergeAndPruneCompletionLists(IEnumerable<CompletionList> completionLists, ICompletionRules completionRules)
+        private static CompletionList MergeAndPruneCompletionLists(IEnumerable<CompletionList> completionLists, CompletionRules completionRules)
         {
             var displayNameToItemsMap = new Dictionary<string, List<CompletionItem>>();
             CompletionItem builder = null;
@@ -227,7 +227,7 @@ namespace Microsoft.CodeAnalysis.Completion
         private static void ReplaceExistingItem(
             CompletionItem item,
             Dictionary<string, List<CompletionItem>> displayNameToItemsMap,
-            ICompletionRules completionRules)
+            CompletionRules completionRules)
         {
             // See if we have an item with 
             var sameNamedItems = displayNameToItemsMap.GetOrAdd(item.DisplayText, s_createList);
@@ -329,7 +329,7 @@ namespace Microsoft.CodeAnalysis.Completion
 
         public abstract Task<TextSpan> GetDefaultTrackingSpanAsync(Document document, int position, CancellationToken cancellationToken);
 
-        public virtual ICompletionRules GetDefaultCompletionRules()
+        public virtual CompletionRules GetDefaultCompletionRules()
         {
             return new CompletionRules(this);
         }
